- 组件库更新。 - 新增拖动条。

+++ src/main/java/github/kasuminova/novaeng/common/hypernet/proc/server/HardwareBandwidthConsumer.java
@@ -2,6 +2,6 @@ package github.kasuminova.novaeng.common.hypernet.proc.server;
 
 public interface HardwareBandwidthConsumer {
 
-    double getHardwareBandwidth();
+    int getHardwareBandwidth();
 
 }
