+++ src/main/java/github/kasuminova/novaeng/common/hypernet/proc/server/HardwareBandwidthProvider.java
@@ -2,6 +2,6 @@ package github.kasuminova.novaeng.common.hypernet.proc.server;
 
 public interface HardwareBandwidthProvider {
 
-    double getHardwareBandwidthProvision();
+    int getHardwareBandwidthProvision();
 
 }
